moved the oracle apply in the oracleDashboard.
